refactor: move assessment items to constants/profile directory
