+++ src/components/profile/FrontendPortfolioView.tsx
@@ -8,7 +8,7 @@ import CheckboxButton from "@/components/ButtonDynamic/CheckboxButton";
 import Modal from "@/components/modal/Modal";
 import type { SelfAssessmentItem } from "./SelfAssessmentGroup";
 import CopyBlackGray from "@/assets/icons/Copy/CopyBlackGray.svg";
-import { FRONTEND_TECH_ITEMS_MAP } from "./constants/frontendAssessmentItems";
+import { FRONTEND_TECH_ITEMS_MAP } from "@/constants/profile/frontendAssessmentItems";
 
 const TECH_ITEMS_MAP = FRONTEND_TECH_ITEMS_MAP;
